requirements.txt
# Add your project dependencies here
pandas
python-dotenv
telethon
psycopg2-binary
dbt-postgres
ultralytics
opencv-python-headless
fastapi
uvicorn[standard]
sqlalchemy